backend/requirements.txt
Flask==2.2.3
flask-cors==5.0.0
Pillow==9.4.0
werkzeug==2.2.2
requests==2.26.0
elasticsearch==7.10.1